+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Minor fix to prevent uneccessary save of device settings
+* (theimo1221) Add Device Setting partial set unit test
 
 ## 1.3.5 (2022-11-05)